+++ __tests__/cadastro.test.tsx
@@ -1,6 +1,7 @@
-import { expect, test, vi } from 'vitest';
+import { assert, expect, test, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/vitest';
+import Cadastro from '@/app/cadastro/page'
 import SubmitForm from '@/app/cadastro/SubmitForm'
 
 vi.mock('next/font/google', () => ({
@@ -16,6 +17,13 @@ vi.mock('next/font/google', () => ({
   }),
 }))
 
+test('Componentes da página de cadastro', async () => {
+  render(<Cadastro />);
+  const main = within(document.querySelector('#cadastro'));
+  expect(
+      main.getByRole('heading', { level: 2, name: 'Cadastro de vagas', exact: true })
+  ).toBeDefined();
+})
 
 test('Formulário de cadastro de estágios', async () => {
   render(<SubmitForm />);
@@ -24,7 +32,7 @@ test('Formulário de cadastro de estágios', async () => {
     'Benefícios', 'Responsabilidades', 'Responsável pelo Estágio',
     'Email do Responsável', 'Como se Candidatar?'
   ].map((label) => (
-    expect(screen.getByLabelText(label)).toBeDefined()
+    expect(screen.getByLabelText(label)).toBeRequired()
   ))
 
   expect(screen.getByRole('button', { name: 'Cadastrar' })).toBeDefined()
